Adding data type resolver for java
